remove code repetition and add additional exception classes. Organize like service.

+++ Service/LikeService.py
@@ -1,13 +1,28 @@
-from Model.model import db, Like, User, Post
+from Exception.exception import LikeAlreadyExistsException, PostNotFoundException, LikeNotFoundException
+from Model.model import db, Like, Post
+
+
+class LikeCountDTO:
+    def __init__(self, like_count):
+        self.like_count = like_count
+
+
+class LikeDTO:
+    def __init__(self, user_id, post_id):
+        self.user_id = user_id
+        self.post_id = post_id
 
 
 class LikeService:
 
     @staticmethod
     def create_like(user_id, post_id):
+        post = Post.query.get(post_id)
+        if not post:
+            raise PostNotFoundException()
         existing_like = Like.query.filter_by(user_id=user_id, post_id=post_id).first()
         if existing_like:
-            return None  # Post already liked
+            raise LikeAlreadyExistsException()
 
         new_like = Like(
             user_id=user_id,
@@ -15,17 +30,24 @@ class LikeService:
         )
         db.session.add(new_like)
         db.session.commit()
-        return new_like
+        return LikeDTO(new_like.user_id, new_like.post_id)
 
     @staticmethod
     def delete_like(user_id, post_id):
+        post = Post.query.get(post_id)
+        if not post:
+            raise PostNotFoundException()
         like = Like.query.filter_by(user_id=user_id, post_id=post_id).first()
         if like:
             db.session.delete(like)
             db.session.commit()
-            return True, "Like deleted successfully"
-        return False, "Failed to delete like"
+            return None
+        raise LikeNotFoundException()
 
     @staticmethod
     def get_post_likes(post_id):
-        return Like.query.filter_by(post_id=post_id).count()
+        post = Post.query.get(post_id)
+        if not post:
+            raise PostNotFoundException()
+        like_count = Like.query.filter_by(post_id=post_id).count()
+        return LikeCountDTO(like_count)
